requirements.txt
asn1crypto==0.24.0
braintree==3.51.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
defusedxml==0.5.0
Django==1.10.4
django-bootstrap-form==3.2.1
django-tinymce==2.4.0
djangorestframework==3.5.3
idna==2.8
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
pycparser==2.19
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
stripe==2.21.0
suds-jurko==0.6
urllib3==1.24.1
xmltodict==0.12.0
whitenoise
